ColonEval/requirements.txt
absl-py==2.3.1
accelerate==1.10.0
annotated-types==0.7.0
anyio==4.9.0
bitsandbytes==0.47.0
certifi==2025.7.9
charset-normalizer==3.4.3
click==8.3.0
deepspeed==0.17.5
distro==1.9.0
einops==0.8.1
exceptiongroup==1.3.0
filelock==3.13.1
fsspec==2024.6.1
h11==0.16.0
hf-xet==1.1.8
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
jinja2==3.1.4
jiter==0.10.0
joblib==1.5.2
markupsafe==2.1.5
mpmath==1.3.0
msgpack==1.1.1
networkx==3.3
ninja==1.13.0
nltk==3.9.2
numpy==2.1.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-ml-py==13.580.65
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
openai==1.93.2
packaging==25.0
pillow==11.3.0
protobuf==6.32.0
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.7
pydantic-core==2.33.2
pyyaml==6.0.2
regex==2025.7.34
requests==2.32.5
rouge-score==0.1.2
safetensors==0.6.2
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
tokenizers==0.21.4
tqdm==4.67.1
transformers==4.55.4
triton==3.1.0
typing-extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0